+++ app/src/main/java/dev/leonlatsch/photok/ui/process/ImportViewModel.kt
@@ -14,7 +14,7 @@
  *   limitations under the License.
  */
 
-package dev.leonlatsch.photok.ui.proccess
+package dev.leonlatsch.photok.ui.process
 
 import android.app.Application
 import android.net.Uri
@@ -24,8 +24,8 @@ import dev.leonlatsch.photok.model.database.entity.Photo
 import dev.leonlatsch.photok.model.database.entity.PhotoType
 import dev.leonlatsch.photok.model.repositories.PhotoRepository
 import dev.leonlatsch.photok.other.getFileName
-import dev.leonlatsch.photok.ui.proccess.base.BaseProcessViewModel
-import dev.leonlatsch.photok.ui.proccess.base.ProcessState
+import dev.leonlatsch.photok.ui.process.base.BaseProcessViewModel
+import dev.leonlatsch.photok.ui.process.base.ProcessState
 import kotlinx.coroutines.launch
 import java.util.*
