+++ app/src/main/java/dev/leonlatsch/photok/ui/process/base/ProcessProgress.kt
@@ -14,11 +14,26 @@
  *   limitations under the License.
  */
 
-package dev.leonlatsch.photok.ui.proccess.base
+package dev.leonlatsch.photok.ui.process.base
 
-enum class ProcessState {
-    INITIALIZE,
-    PROCESSING,
-    FINISHED,
-    ABORTED
+import androidx.lifecycle.MutableLiveData
+
+class ProcessProgress {
+
+    val maxPercent: Int = 100
+    val progressPercent: MutableLiveData<Int> = MutableLiveData()
+
+    val max: MutableLiveData<Int> = MutableLiveData()
+    val current: MutableLiveData<Int> = MutableLiveData()
+
+    init {
+        max.postValue(0)
+        current.postValue(0)
+    }
+
+    fun update(current: Int, from: Int) {
+        this.max.postValue(from)
+        this.current.postValue(current)
+        progressPercent.postValue((current * maxPercent) / from)
+    }
 }
